apps/docusign/src/inngest/middlewares/rate-limit-middleware.test.ts
import { afterAll, beforeAll, describe, expect, test, vi } from 'vitest';
import { RetryAfterError } from 'inngest';
import { DocusignError } from '@/connectors/common/error';
import { rateLimitMiddleware } from './rate-limit-middleware';

const rateLimitReset = 1717515430; // Unix timestamp in seconds

describe('rate-limit middleware', () => {
  beforeAll(() => {
    vi.setSystemTime(new Date((rateLimitReset - 30) * 1000));
  });

  afterAll(() => {
    vi.useRealTimers();
  });

  test('should not transform the output when their is no error', () => {
    expect(
      rateLimitMiddleware
        .init()
        // @ts-expect-error -- this is a mock
        .onFunctionRun({ fn: { name: 'foo' } })
        .transformOutput({
          result: {},
        })
    ).toBeUndefined();
  });

  test('should not transform the output when the error is not about Docusign rate limit', () => {
    expect(
      rateLimitMiddleware
        .init()
        // @ts-expect-error -- this is a mock
        .onFunctionRun({ fn: { name: 'foo' } })
        .transformOutput({
          result: {
            error: new Error('foo bar'),
          },
        })
    ).toBeUndefined();
  });

  test('should transform the output error to RetryAfterError when the error is about Docusign rate limit', () => {
    const rateLimitError = new DocusignError('foo bar', {
      // @ts-expect-error this is a mock
      response: {
        status: 429,
        headers: new Headers({
          'X-RateLimit-Limit': '100',
          'X-RateLimit-Remaining': '0',
          'X-RateLimit-Reset': `${rateLimitReset}`,
        }),
      },
    });

    const context = {
      foo: 'bar',
      baz: {
        biz: true,
      },
      result: {
        data: 'bizz',
        error: rateLimitError,
      },
    };

    const result = rateLimitMiddleware
      .init()
      // @ts-expect-error -- this is a mock
      .onFunctionRun({ fn: { name: 'foo' } })
      .transformOutput(context);
    expect(result?.result.error).toBeInstanceOf(RetryAfterError);
    expect(result?.result.error.retryAfter).toStrictEqual('30');
    expect(result).toMatchObject({
      foo: 'bar',
      baz: {
        biz: true,
      },
      result: {
        data: 'bizz',
      },
    });
  });
});
